[Reflection] Make reflection expect default inline class parameters with nullable field boxed
The proper solution would be to fix the cause:
```kotlin
inline class IC(val value: Int?)

fun f(x: IC = IC(2))
```
IR sets type of `x` to `IC?` as default value of a class in JVM is null.
It is not possible to distinguish `null: IC?` and `IC(null): IC?` without boxing.
Even though it is redundant here and should be removed,
it is impossible to do it due to breakage of the ABI.

#KT-57357