[K/N][test] Test execution of Kotlin/Native tests using XCTest
Make "native.tests" infrastructure be able to run xctest-tests on a
local machine. The XCTestExecutor adds an ability to run directly on
a host target or in a simulator environment.

A test infrastructure gets a new XCTEST_FRAMEWORK property to specify
a location of the XCTest.framework. A kotlin-test-native-xctest library
passed as CUSTOM_KLIBS to the infrastructure.

Part of ^KT-58928