[klib] Fix exception for clashing signatures from different modules
If we encounter a declaration in the current module whose signature
is the same as that of a declaration in another module which we happen
to also reference from the current module, report the error only
on the declaration in the current module. The declarations from other
modules may not have an `IrFile` which is necessary for showing an error
message.

^KT-65063 Fixed