Fix K2 CodeGen annotation parameter resolution failure
When we have parameters of annotation like
```
// FILE: Anno.kt
package p3

@Target(AnnotationTarget.FUNCTION)
annotation class Anno(vararg val x: String)

// FILE: main.kt
import p3.Anno

@Anno("A", "B")
fun foo(): Int = 10
```
the K2 CodeGen causes an exception reporting the unresolved type
reference. A lazy resolution call fixes the issue.

In addition, FIR for the value parameter to IR symbol conversion fails
because the initializer of the constructor parameter is fake. The fake
parameter is the same as an access to the FIR property for the
constructor parameter. It is created by `buildPropertyAccessExpression`
in `KtParameter.toFirProperty()`. Since the FIR2IR does not put the
IrValueParameterSymbol to `localStorage` when it is created by
`buildPropertyAccessExpression`, the FIR2IR for the parameter
initializer fails on getting `IrValueParameterSymbol` from
`localStorage`. To fix this issue, this commit supports
`IrValueParameterSymbol` creation for a given `FirValueParameter` when
`localStorage` does not have it.

^KT-65099 Fixed